Rename: 컨테이너리스트와 프로필페이지 내 모든 컴포넌트 요소 알맞은 폴더 생성 후 이동
